move state management code out of source-relational-db
state management code is not specific to relational databases so moving this to its own module makes sense
