+++ airbyte-integrations/connectors/source-db/.dockerignore
@@ -1,0 +1,3 @@
+*
+!Dockerfile
+!build
