+++ airbyte-integrations/connectors/source-db/Dockerfile
@@ -1,0 +1,20 @@
+FROM airbyte/integration-base-java:dev AS build
+
+WORKDIR /airbyte
+
+ENV APPLICATION source-db
+
+COPY build/distributions/${APPLICATION}*.tar ${APPLICATION}.tar
+
+RUN tar xf ${APPLICATION}.tar --strip-components=1 && rm -rf ${APPLICATION}.tar
+
+FROM airbyte/integration-base-java:dev
+
+WORKDIR /airbyte
+
+ENV APPLICATION source-db
+
+COPY --from=build /airbyte /airbyte
+
+LABEL io.airbyte.version=0.3.1
+LABEL io.airbyte.name=airbyte/source-db
