+++ airbyte-integrations/connectors/source-db/README.md
@@ -1,0 +1,3 @@
+# Database Source
+
+We are not planning to expose this source in the UI. It serves as a base upon which we can build all of our other Database sources.
