+++ airbyte-integrations/connectors/source-db/build.gradle
@@ -1,0 +1,33 @@
+import org.jsonschema2pojo.SourceType
+
+plugins {
+    id "java-library"
+    id "java-test-fixtures"
+    id "com.github.eirnym.js2p" version "1.0"
+}
+
+dependencies {
+    implementation libs.airbyte.protocol
+    implementation project(':airbyte-commons')
+    implementation project(':airbyte-config-oss:config-models-oss')
+    implementation project(':airbyte-db:db-lib')
+
+    testImplementation libs.connectors.testcontainers
+
+    testFixturesImplementation libs.airbyte.protocol
+    testFixturesImplementation libs.connectors.testcontainers
+}
+
+jsonSchema2Pojo {
+    sourceType = SourceType.YAMLSCHEMA
+    source = files("${sourceSets.main.output.resourcesDir}/db_models")
+    targetDirectory = new File(project.buildDir, 'generated/src/gen/java/')
+    removeOldOutput = true
+
+    targetPackage = 'io.airbyte.integrations.source.db.models'
+
+    useLongIntegers = true
+    generateBuilders = true
+    includeConstructors = false
+    includeSetters = true
+}
