+++ airbyte-integrations/connectors/source-db/src/main/java/io/airbyte/integrations/source/db/state/CursorInfo.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2023 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.source.relationaldb;
+package io.airbyte.integrations.source.db.state;
 
 import java.util.Objects;
