+++ airbyte-integrations/connectors/source-db/src/main/java/io/airbyte/integrations/source/db/state/CursorManager.java
@@ -2,10 +2,9 @@
  * Copyright (c) 2023 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.source.relationaldb.state;
+package io.airbyte.integrations.source.db.state;
 
 import com.google.common.annotations.VisibleForTesting;
-import io.airbyte.integrations.source.relationaldb.CursorInfo;
 import io.airbyte.protocol.models.v0.AirbyteStreamNameNamespacePair;
 import io.airbyte.protocol.models.v0.ConfiguredAirbyteCatalog;
 import io.airbyte.protocol.models.v0.ConfiguredAirbyteStream;
