+++ airbyte-integrations/connectors/source-db/src/main/resources/db_models/db_models.yaml
