+++ airbyte-integrations/connectors/source-relational-db/build.gradle
@@ -2,6 +2,7 @@ import org.jsonschema2pojo.SourceType
 
 plugins {
     id "java-library"
+    id "java-test-fixtures"
     id "com.github.eirnym.js2p" version "1.0"
 }
 
@@ -9,6 +10,7 @@ dependencies {
     implementation project(':airbyte-commons')
     implementation project(':airbyte-db:db-lib')
     implementation project(':airbyte-integrations:bases:base-java')
+    implementation project(':airbyte-integrations:connectors:source-db')
     implementation libs.airbyte.protocol
     implementation project(':airbyte-json-validation')
     implementation project(':airbyte-config-oss:config-models-oss')
@@ -24,6 +26,7 @@ dependencies {
     testImplementation libs.junit.jupiter.system.stubs
 
     implementation files(project(':airbyte-integrations:bases:base-java').airbyteDocker.outputs)
+    testImplementation testFixtures(project(':airbyte-integrations:connectors:source-db'))
 }
 
 jsonSchema2Pojo {
@@ -32,7 +35,7 @@ jsonSchema2Pojo {
     targetDirectory = new File(project.buildDir, 'generated/src/gen/java/')
     removeOldOutput = true
 
-    targetPackage = 'io.airbyte.integrations.source.relationaldb.models'
+    targetPackage = 'io.airbyte.integrations.source.db.models'
 
     useLongIntegers = true
     generateBuilders = true
