+++ airbyte-integrations/connectors/source-relational-db/src/main/java/io/airbyte/integrations/source/relationaldb/state/GlobalStateManager.java
@@ -4,16 +4,19 @@
 
 package io.airbyte.integrations.source.relationaldb.state;
 
-import static io.airbyte.integrations.source.relationaldb.state.StateGeneratorUtils.CURSOR_FIELD_FUNCTION;
-import static io.airbyte.integrations.source.relationaldb.state.StateGeneratorUtils.CURSOR_FUNCTION;
-import static io.airbyte.integrations.source.relationaldb.state.StateGeneratorUtils.CURSOR_RECORD_COUNT_FUNCTION;
-import static io.airbyte.integrations.source.relationaldb.state.StateGeneratorUtils.NAME_NAMESPACE_PAIR_FUNCTION;
+import static io.airbyte.integrations.source.db.state.StateGeneratorUtils.CURSOR_FIELD_FUNCTION;
+import static io.airbyte.integrations.source.db.state.StateGeneratorUtils.CURSOR_FUNCTION;
+import static io.airbyte.integrations.source.db.state.StateGeneratorUtils.CURSOR_RECORD_COUNT_FUNCTION;
+import static io.airbyte.integrations.source.db.state.StateGeneratorUtils.NAME_NAMESPACE_PAIR_FUNCTION;
 
 import io.airbyte.commons.json.Jsons;
-import io.airbyte.integrations.source.relationaldb.CdcStateManager;
-import io.airbyte.integrations.source.relationaldb.models.CdcState;
-import io.airbyte.integrations.source.relationaldb.models.DbState;
-import io.airbyte.integrations.source.relationaldb.models.DbStreamState;
+import io.airbyte.integrations.source.db.state.AbstractStateManager;
+import io.airbyte.integrations.source.db.state.CdcStateManager;
+import io.airbyte.integrations.source.db.models.CdcState;
+import io.airbyte.integrations.source.db.models.DbState;
+import io.airbyte.integrations.source.db.models.DbStreamState;
+import io.airbyte.integrations.source.db.state.StateGeneratorUtils;
+import io.airbyte.integrations.source.db.state.StateManager;
 import io.airbyte.protocol.models.v0.AirbyteGlobalState;
 import io.airbyte.protocol.models.v0.AirbyteStateMessage;
 import io.airbyte.protocol.models.v0.AirbyteStateMessage.AirbyteStateType;
