+++ airbyte-integrations/connectors/source-relational-db/src/main/java/io/airbyte/integrations/source/relationaldb/state/StateManagerFactory.java
@@ -5,7 +5,10 @@
 package io.airbyte.integrations.source.relationaldb.state;
 
 import io.airbyte.commons.json.Jsons;
-import io.airbyte.integrations.source.relationaldb.models.DbState;
+import io.airbyte.integrations.source.db.state.StateGeneratorUtils;
+import io.airbyte.integrations.source.db.state.StateManager;
+import io.airbyte.integrations.source.db.state.StreamStateManager;
+import io.airbyte.integrations.source.db.models.DbState;
 import io.airbyte.protocol.models.v0.AirbyteStateMessage;
 import io.airbyte.protocol.models.v0.AirbyteStateMessage.AirbyteStateType;
 import io.airbyte.protocol.models.v0.ConfiguredAirbyteCatalog;
