+++ airbyte-integrations/connectors/source-relational-db/src/test/java/io/airbyte/integrations/source/relationaldb/AbstractDbSourceTest.java
@@ -12,7 +12,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import io.airbyte.commons.features.EnvVariableFeatureFlags;
 import io.airbyte.commons.json.Jsons;
 import io.airbyte.commons.resources.MoreResources;
-import io.airbyte.integrations.source.relationaldb.state.StateGeneratorUtils;
+import io.airbyte.integrations.source.db.state.StateGeneratorUtils;
 import io.airbyte.protocol.models.v0.AirbyteStateMessage;
 import io.airbyte.protocol.models.v0.AirbyteStateMessage.AirbyteStateType;
 import java.io.IOException;
