+++ airbyte-integrations/connectors/source-relational-db/src/test/java/io/airbyte/integrations/source/relationaldb/state/GlobalStateManagerTest.java
@@ -4,22 +4,24 @@
 
 package io.airbyte.integrations.source.relationaldb.state;
 
-import static io.airbyte.integrations.source.relationaldb.state.StateTestConstants.CURSOR_FIELD1;
-import static io.airbyte.integrations.source.relationaldb.state.StateTestConstants.CURSOR_FIELD2;
-import static io.airbyte.integrations.source.relationaldb.state.StateTestConstants.NAMESPACE;
-import static io.airbyte.integrations.source.relationaldb.state.StateTestConstants.NAME_NAMESPACE_PAIR1;
-import static io.airbyte.integrations.source.relationaldb.state.StateTestConstants.STREAM_NAME1;
-import static io.airbyte.integrations.source.relationaldb.state.StateTestConstants.STREAM_NAME2;
-import static io.airbyte.integrations.source.relationaldb.state.StateTestConstants.STREAM_NAME3;
+import static io.airbyte.integrations.source.db.state.StateTestConstants.CURSOR_FIELD1;
+import static io.airbyte.integrations.source.db.state.StateTestConstants.CURSOR_FIELD2;
+import static io.airbyte.integrations.source.db.state.StateTestConstants.NAMESPACE;
+import static io.airbyte.integrations.source.db.state.StateTestConstants.NAME_NAMESPACE_PAIR1;
+import static io.airbyte.integrations.source.db.state.StateTestConstants.STREAM_NAME1;
+import static io.airbyte.integrations.source.db.state.StateTestConstants.STREAM_NAME2;
+import static io.airbyte.integrations.source.db.state.StateTestConstants.STREAM_NAME3;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
 
 import io.airbyte.commons.json.Jsons;
-import io.airbyte.integrations.source.relationaldb.models.CdcState;
-import io.airbyte.integrations.source.relationaldb.models.DbState;
-import io.airbyte.integrations.source.relationaldb.models.DbStreamState;
+import io.airbyte.integrations.source.db.models.CdcState;
+import io.airbyte.integrations.source.db.models.DbState;
+import io.airbyte.integrations.source.db.models.DbStreamState;
+import io.airbyte.integrations.source.db.state.StateManager;
 import io.airbyte.protocol.models.v0.AirbyteGlobalState;
 import io.airbyte.protocol.models.v0.AirbyteStateMessage;
 import io.airbyte.protocol.models.v0.AirbyteStateMessage.AirbyteStateType;
