+++ airbyte-integrations/connectors/source-relational-db/src/test/java/io/airbyte/integrations/source/relationaldb/state/StateGeneratorUtilsTest.java
@@ -7,6 +7,7 @@ package io.airbyte.integrations.source.relationaldb.state;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import io.airbyte.integrations.source.db.state.StateGeneratorUtils;
 import io.airbyte.protocol.models.v0.StreamDescriptor;
 import org.junit.jupiter.api.Test;
